# api/openapi.yaml
openapi: 3.1.0
info:
  title: LXC File Hub API
  version: 1.0.0
  description: REST API for LXC File Hub (lxcfh) — section 10.2
servers:
  - url: /api/v1
    description: API v1 base path
paths:
...
  /nodes:
    get:
      tags: [Nodes]
      summary: List nodes
      operationId: listNodes
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Node list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Node'
    post:
      tags: [Nodes]
      summary: Create node
      operationId: createNode
      security:
        - sessionAuth: []
        - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Node'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Node'
  /nodes/{nodeID}:
    parameters:
      - $ref: '#/components/parameters/nodeID'
    get:
      tags: [Nodes]
      summary: Get node
      operationId: getNode
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Node
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Node'
    patch:
      tags: [Nodes]
      summary: Update node
      operationId: updateNode
      security:
        - sessionAuth: []
        - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Node'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Node'
    delete:
      tags: [Nodes]
      summary: Delete node
      operationId: deleteNode
      security:
        - sessionAuth: []
        - csrfToken: []
      responses:
        '204':
          description: Deleted
...
  /nodes/{nodeID}/test:
    parameters:
      - $ref: '#/components/parameters/nodeID'
    post:
      tags: [Nodes]
      summary: Test node connection
      operationId: testNode
      security:
        - sessionAuth: []
        - csrfToken: []
      responses:
        '200':
          description: Test result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestResult'
  /nodes/{nodeID}/accept-host-key:
    parameters:
      - $ref: '#/components/parameters/nodeID'
    post:
      tags: [Nodes]
      summary: Accept SSH host key
      operationId: acceptHostKey
      security:
        - sessionAuth: []
        - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fingerprint]
              properties:
                fingerprint:
                  type: string
      responses:
        '200':
          description: Updated node
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Node'
...
  /transfers:
    get:
      tags: [Transfers]
      summary: List transfers
      operationId: listTransfers
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Transfer list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transfer'
    post:
      tags: [Transfers]
      summary: Create transfer
      operationId: createTransfer
      security:
        - sessionAuth: []
        - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transfer'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'
  /transfers/{transferID}:
    parameters:
      - $ref: '#/components/parameters/transferID'
    get:
      tags: [Transfers]
      summary: Get transfer
      operationId: getTransfer
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Transfer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'
    delete:
      tags: [Transfers]
      summary: Delete transfer
      operationId: deleteTransfer
      security:
        - sessionAuth: []
        - csrfToken: []
      responses:
        '204':
          description: Deleted
  /transfers/{transferID}/pause:
    parameters:
      - $ref: '#/components/parameters/transferID'
    post:
      tags: [Transfers]
      summary: Pause transfer
      operationId: pauseTransfer
      security:
        - sessionAuth: []
        - csrfToken: []
      responses:
        '200':
          description: Paused
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'
  /transfers/{transferID}/resume:
    parameters:
      - $ref: '#/components/parameters/transferID'
    post:
      tags: [Transfers]
      summary: Resume transfer
      operationId: resumeTransfer
      security:
        - sessionAuth: []
        - csrfToken: []
      responses:
        '200':
          description: Resumed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'
  /transfers/{transferID}/cancel:
    parameters:
      - $ref: '#/components/parameters/transferID'
    post:
      tags: [Transfers]
      summary: Cancel transfer
      operationId: cancelTransfer
      security:
        - sessionAuth: []
        - csrfToken: []
      responses:
        '200':
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'
  /transfers/{transferID}/retry:
    parameters:
      - $ref: '#/components/parameters/transferID'
    post:
      tags: [Transfers]
      summary: Retry transfer
      operationId: retryTransfer
      security:
        - sessionAuth: []
        - csrfToken: []
      responses:
        '200':
          description: Retried
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'
  /uploads:
    post:
      tags: [Uploads]
      summary: Create tus upload
      operationId: createUpload
      security:
        - sessionAuth: []
        - csrfToken: []
      parameters:
        - name: Upload-Length
          in: header
          schema:
            type: integer
        - name: Upload-Node-ID
          in: header
          schema:
            type: string
        - name: Upload-Path
          in: header
          schema:
            type: string
      responses:
        '201':
          description: Upload created
          headers:
            Location:
              schema:
                type: string
            Upload-Offset:
              schema:
                type: integer
            Tus-Resumable:
              schema:
                type: string
  /uploads/{uploadID}:
    parameters:
      - $ref: '#/components/parameters/uploadID'
    head:
      tags: [Uploads]
      summary: tus HEAD upload
      operationId: headUpload
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Upload status
          headers:
            Upload-Offset:
              schema:
                type: integer
            Upload-Length:
              schema:
                type: integer
            Tus-Resumable:
              schema:
                type: string
    patch:
      tags: [Uploads]
      summary: tus PATCH upload
      operationId: patchUpload
      security:
        - sessionAuth: []
        - csrfToken: []
      parameters:
        - name: Upload-Offset
          in: header
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/offset+octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '204':
          description: Chunk uploaded
          headers:
            Upload-Offset:
              schema:
                type: integer
            Tus-Resumable:
              schema:
                type: string
    get:
      tags: [Uploads]
      summary: Get upload metadata
      operationId: getUpload
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Upload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Upload'
    delete:
      tags: [Uploads]
      summary: Delete upload
      operationId: deleteUpload
      security:
        - sessionAuth: []
        - csrfToken: []
      responses:
        '204':
          description: Deleted
  /config/export:
    get:
      tags: [Config]
      summary: Export configuration as YAML
      operationId: configExport
      security:
        - sessionAuth: []
      responses:
        '200':
          description: YAML configuration
          content:
            application/x-yaml:
              schema:
                type: string
  /config/validate:
    post:
      tags: [Config]
      summary: Validate configuration YAML
      operationId: configValidate
      security:
        - sessionAuth: []
        - csrfToken: []
      requestBody:
        required: true
        content:
          application/x-yaml:
            schema:
              type: string
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResult'
  /config/preview:
    post:
      tags: [Config]
      summary: Preview configuration changes
      operationId: configPreview
      security:
        - sessionAuth: []
        - csrfToken: []
      requestBody:
        required: true
        content:
          application/x-yaml:
            schema:
              type: string
      responses:
        '200':
          description: Preview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigPreview'
  /config/apply:
    post:
      tags: [Config]
      summary: Apply configuration (full replacement)
      operationId: configApply
      security:
        - sessionAuth: []
        - csrfToken: []
      requestBody:
        required: true
        content:
          application/x-yaml:
            schema:
              type: string
      responses:
        '204':
          description: Applied
  /settings:
    get:
      tags: [Settings]
      summary: Get settings
      operationId: getSettings
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
    patch:
      tags: [Settings]
      summary: Patch settings
      operationId: patchSettings
      security:
        - sessionAuth: []
        - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Settings'
      responses:
        '200':
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: lxcfh_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  parameters:
    nodeID:
      name: nodeID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    transferID:
      name: transferID
      in: path
      required: true
      schema:
        type: string
        format: uuid
    uploadID:
      name: uploadID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Node:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        host:
          type: string
        port:
          type: integer
        username:
          type: string
        credential_id:
          type: string
        key_id:
          type: string
        labels:
          type: object
          additionalProperties:
            type: string
        enabled:
          type: boolean
        host_key_status:
          type: string
        host_key_fingerprint:
          type: string
        last_ping_at:
          type: string
          format: date-time
        last_ping_status:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    TestResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
...
    Transfer:
      type: object
      properties:
        id:
          type: string
        node_id:
          type: string
        source_path:
          type: string
        dest_path:
          type: string
        direction:
          type: string
        status:
          type: string
          enum: [pending, running, paused, completed, failed, cancelled]
        bytes_total:
          type: integer
        bytes_done:
          type: integer
        error:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
    Upload:
      type: object
      properties:
        id:
          type: string
        node_id:
          type: string
        path:
          type: string
        size:
          type: integer
        offset:
          type: integer
        status:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Settings:
      type: object
      properties:
        site_name:
          type: string
        session_timeout_min:
          type: integer
        max_upload_size_mb:
          type: integer
        rate_limit_per_minute:
          type: integer
        require_reauth:
          type: boolean
        allow_registration:
          type: boolean
        default_node_port:
          type: integer
        backup_retention_days:
          type: integer
    ValidationResult:
      type: object
      properties:
        valid:
          type: boolean
        errors:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
    ConfigPreview:
      type: object
      properties:
        changes:
          type: array
          items:
            $ref: '#/components/schemas/ConfigChange'
    ConfigChange:
      type: object
      properties:
        resource:
          type: string
        action:
          type: string
        detail:
          type: string
...